Updates lib functions to ignore volume changes
Earlier implementation of update_volume_isothermal() was making use of
new volume as input till which the gas expands.
But this assumption wasn't making sense, I updated the client code send
NAN and ignore it on the library side if NAN is received.
This ignoring part of code become a dead code and hence not testable or
the code coverage got reduced.

Fix: the current code change will ignore volume argument of
update_volume_isothermal() and hence dead code part can be removed.